requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-paypal==2.1
djangorestframework==3.15.2
idna==3.10
paypal==1.2.5
paypalrestsdk==1.13.3
pillow==11.0.0
pip==24.2
pycparser==2.22
pyOpenSSL==24.2.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
urllib3==2.2.3